wrequirements.txt
python-dotenv
undetected-chromedriver
selenium
windows-curses
